修改部分bug

+++ app/LKT/webapp/modules/api/actions/addressAction.class.php
@@ -87,8 +87,6 @@ class addressAction extends Action {
             echo json_encode(array('status'=>0,'err'=>'删除失败!'));
             exit();
         }
-        // $user_id = $r[0]->user_id;
-        // echo json_encode(array('openid' => $openid,'arr'=>$arr ));
     }
     // 设置默认
     public function set_default(){
@@ -100,13 +98,16 @@ class addressAction extends Action {
         // 根据微信id,查询用户id
         $sql = "select * from lkt_user where wx_id = '$openid'";
         $r = $db->select($sql);
-        $user_id = $r[0]->user_id;
+        if($r){
+            $user_id = $r[0]->user_id;
+            $sql = "update lkt_user_address set is_default = 0 where uid = '$user_id'";
+            $r = $db->update($sql);
+            $sql = "update lkt_user_address set is_default = 1 where uid = '$user_id' and id = '$addr_id'";
+            $rr = $db->update($sql);
+        }else{
+            $rr = 0;
+        }
 
-        $sql = "update lkt_user_address set is_default = 0 where uid = '$user_id'";
-        $r = $db->update($sql);
-
-        $sql = "update lkt_user_address set is_default = 1 where uid = '$user_id' and id = '$addr_id'";
-        $rr = $db->update($sql);
         if($rr > 0){
             echo json_encode(array('status'=>1,'err'=>'操作成功!'));
             exit();
@@ -127,17 +128,23 @@ class addressAction extends Action {
         // 根据微信id,查询用户id
         $sql = "select * from lkt_user where wx_id = '$openid'";
         $r = $db->select($sql);
-        $user_id = $r[0]->user_id;
-        // 根据用户id,查询地址表
-        $sql = "delete from lkt_user_address where uid = '$user_id' and id = '$id_arr'";
-        $r = $db->delete($sql);
-        if($r > 0){
-            echo json_encode(array('status'=>1));
-            exit();
+        if($r){
+            $user_id = $r[0]->user_id;
+            // 根据用户id,查询地址表
+            $sql = "delete from lkt_user_address where uid = '$user_id' and id = '$id_arr'";
+            $r = $db->delete($sql);
+            if($r > 0){
+                echo json_encode(array('status'=>1));
+                exit();
+            }else{
+                echo json_encode(array('status'=>0,'err'=>'删除失败'));
+                exit();
+            }
         }else{
             echo json_encode(array('status'=>0,'err'=>'删除失败'));
             exit();
         }
+
         
         return;
     }
@@ -200,25 +207,46 @@ class addressAction extends Action {
         $id_arr = $_POST['id_arr']; // 地址id
         $sql = "select * from lkt_user_address where id = '$id_arr'"; //查询修改前的详细地址
         $r = $db->select($sql);
-        $sheng = $r[0]->sheng;//省
-        $city = $r[0]->city;//市
-        $quyu = $r[0]->quyu;//县
+        if($r){
+            $sheng = $r[0]->sheng;//省
+            $city = $r[0]->city;//市
+            $quyu = $r[0]->quyu;//县
 
-        // 查询省的编号
-        $sql01 = "select G_CName from admin_cg_group a where a.GroupID='$sheng'";
-        $r01 = $db->select($sql01);
-        $province = $r01[0]->G_CName;
-        // 根据省查询市
-        $sql02 = "select G_CName from admin_cg_group a where a.GroupID='$city'";
-        $r02 = $db->select($sql02);
-        $city = $r02[0]->G_CName;
+            // 查询省的编号
+            $sql01 = "select G_CName from admin_cg_group a where a.GroupID='$sheng'";
+            $r01 = $db->select($sql01);
+            if($r01){
+                $province = $r01[0]->G_CName;
+            }else{
+                $province = '';
+            }
+            $province = $r01[0]->G_CName;
+            // 根据省查询市
+            $sql02 = "select G_CName from admin_cg_group a where a.GroupID='$city'";
+            $r02 = $db->select($sql02);
+            if($r02){
+                $city = $r02[0]->G_CName;
+            }else{
+                $city = '';
+            }
 
-        // 根据市查询县
-        $sql03 = "select G_CName from admin_cg_group a where a.GroupID='$quyu'";
-        $r03 = $db->select($sql03);
-        $county = $r03[0]->G_CName;
-        echo json_encode(array('adds'=>$r,'province'=>$province,'city'=>$city,'county'=>$county));
-        exit();
+            
+            // 根据市查询县
+            $sql03 = "select G_CName from admin_cg_group a where a.GroupID='$quyu'";
+            $r03 = $db->select($sql03);
+            if($r03){
+                $county = $r03[0]->G_CName;
+            }else{
+                $county = '';
+            }
+            
+
+            echo json_encode(array('adds'=>$r,'province'=>$province,'city'=>$city,'county'=>$county));
+            exit();
+        }else{
+            echo json_encode(array('status'=>0,'info'=>'手机号码有误！'));
+        }
+
         return;
     }
 }
